Removed prop passing, but lots of duplicate code
