--- src/main/css/components/timeline/HomeTimeline.scss
@@ -1,19 +1,0 @@
-.home-timeline {
-	display:inline-block;
-	vertical-align: top;
-	text-align: left;
-
-	.timeline-div {
-		.tweets {
-			.row {
-				&:nth-child(odd) {
-					background-color:#e8f5fd;
-				}
-
-				&:nth-child(even) {
-					background-color:#e9e9e9;
-				}
-			}
-		}
-	}
-}
