--- src/main/css/components/timeline/TimelineUI.scss
@@ -1,29 +1,0 @@
-// Code in common between HomeTimeline and UserTimeline
-.timeline-div {
-
-	.button-div {
-		.update-timeline {
-			margin-bottom: 4px;
-			text-align: left;
-		}
-	}
-
-	.error-div {
-		font-size: 12px;
-		margin-top: 5px;
-	}
-
-	.tweets {
-		border: 2px solid;
-		margin-top: 10px;
-
-		.row {
-			border-bottom: 1px solid #bbbbbb;
-
-			&:last-child {
-				border-bottom: none;
-			}
-		}
-	}
-
-}
