--- src/main/css/components/timeline/UserTimeline.scss
@@ -1,25 +1,0 @@
-.user-timeline {
-	display:inline-block;
-	vertical-align: top;
-	text-align: left;
-
-	.timeline-div {
-
-		.button-div {
-			display: flex;
-			justify-content: flex-end;
-		}
-
-		.tweets {
-			.row {
-				&:nth-child(odd) {
-					background-color:#f5f8fa;
-				}
-
-				&:nth-child(even) {
-					background-color:#1da1f2;
-				}
-			}
-		}
-	}
-}
