+++ src/main/js/components/timeline/UserTimelineUI.js
@@ -1,24 +1,81 @@
 import React, { Component } from "react";
-import TimelineUI from "./TimelineUI";
+import TweetBlock from "./TweetBlock";
+import "../../../css/components/timeline/UserTimelineUI.scss";
 import { fetchUserTimeline } from "../../twitter-api";
-import "../../../css/components/timeline/UserTimeline.scss";
 
 export default class UserTimelineUI extends React.Component {
 
-	static get buttonName() {
-		return "Update User Timeline";
+	constructor() {
+		super();
+		this.state = {
+			tweets: null,
+			message: null
+		}
+		this.updateTimeline = this.updateTimeline.bind(this);
 	}
 
-	static get title() {
-		return "User Timeline";
+	static get apiErrorMessage() {
+		return "Failed to fetch tweets from user timeline. Please try again later.";
+	}
+
+	static get emptyTimelineMessage() {
+		return "No tweets are available, post a tweet!";
+	}
+
+	static formatTweet(tweet) {
+		tweet.createdAt = new Date(tweet.createdAt).toLocaleString("en-us", {month: "short", day: "numeric"});
+		if (!tweet.user) {
+			tweet.user = {
+				name: "Unknown User"
+			};
+		} else {
+			tweet.twitterHandle = "";
+		}
+		return tweet;
+	}
+
+	componentDidMount() {
+		this.updateTimeline();
 	}
 
 	render() {
+		let displayedElem; // Either display message or tweets
+		if (this.state.message) {
+			displayedElem = <div className="error-div">{this.state.message}</div>
+		}
+		if (this.state.tweets) {
+			// Map tweets to React Components
+			const tweets = this.state.tweets.map(tweet => {
+				return (
+					<div key={tweet.url} className="row">
+						<TweetBlock tweet={UserTimelineUI.formatTweet(tweet)}/>
+					</div>
+				);
+			});
+			displayedElem = <div className="tweets">{tweets}</div>;
+		} 
 		return (
 			<div className="user-timeline timeline-comp">
-				<TimelineUI apiCall={fetchUserTimeline} buttonName={UserTimelineUI.buttonName} 
-						title={UserTimelineUI.title} hideHandle={true}/>
+				<h3 className="title">User Timeline</h3>
+				<div className="button-div">
+					<button className="update-timeline" onClick={this.updateTimeline}>Update User Timeline</button>
+				</div>
+				{displayedElem}
 			</div>
 		);
+	}
+
+	updateTimeline() {
+		fetchUserTimeline((err, tweets) => {
+			if (err) {
+				this.setState({tweets: null, message: UserTimelineUI.apiErrorMessage});
+			} else {
+				if (tweets.length) {
+					this.setState({tweets: tweets, message: null});
+				} else {
+					this.setState({tweets : null, message: UserTimelineUI.emptyTimelineMessage});
+				}
+			}
+		});
 	}
 }
